machineLearning/api/socketServer/requirements.txt
pymongo
scikit-learn
numpy
scipy
schedule
requests==2.18.4
redis==2.10.5
python-dotenv==0.9.1
